requirements.txt
streamlit
pymupdf
python-docx
pdf2image
pytesseract
rake-nltk
spacy
nltk
